+++ CHANGES.md
@@ -6,6 +6,10 @@ Changes in v1.1b1
 
 - Added `PAPPL_SOPTIONS_NO_TLS` option to disable TLS support.
 - Added Wi-Fi callbacks to support configuration over IPP-USB (Issue #45)
+- `papplMainLoop` now uses a persistent location for state and spool files by
+  default (Issue #128)
+- `papplMainLoop` now supports clients talking to a system-wide server running
+  as root (Issue #148)
 
 
 Changes in v1.0.3